more move cleanup and improvements
