+++ packages/ignite-basic-structure/templates/Unholy/Containers/Styles/LaunchScreenStyle.js
