+++ packages/ignite-dev-screens/index.js
@@ -3,7 +3,7 @@ const copyDevScreens = async function (context) {
   const { parameters, print, strings, ignite } = context
   const { isBlank } = strings
 
-  const screens = ["AllComponentsScreen", "APITestingScreen", "DeviceInfoScreen", "PresentationScreen", "ThemeScreen", "PluginExamplesScreen", "LaunchScreen"]
+  const screens = ["APITestingScreen", "ComponentExamplesScreen", "DeviceInfoScreen", "DevscreensButton", "PluginExamplesScreen", "PresentationScreen", "ThemeScreen"]
 
   const jobs = []
 
@@ -19,17 +19,6 @@ const copyDevScreens = async function (context) {
       }
     )
   })
-
-  jobs.push(
-    {
-      template: 'DevscreensButton.ejs',
-      target: 'ignite/DevScreens/DevscreensButton.js'
-    },
-    {
-      template: 'DevscreensButtonStyle.ejs',
-      target: 'ignite/DevScreens/Styles/DevscreensButtonStyle.js'
-    }
-  )
 
   // make the templates
   await ignite.copyBatch(context, jobs, {})
@@ -53,7 +42,7 @@ const add = async function (context) {
   const launchScreen = `${process.cwd()}/App/Containers/LaunchScreen.js`
   if (filesystem.exists(launchScreen)) {
     if (!patching.isInFile(launchScreen, 'import DevscreensButton')) {
-      patching.insertInFile(launchScreen, 'import styles', 'import DevscreensButton from \'./Components/DevscreensButton.js\'\n', false)
+      patching.insertInFile(launchScreen, 'import styles', 'import DevscreensButton from \'../../ignite/DevScreens/DevscreensButton.js\'\n', false)
     }
     if (!patching.isInFile(launchScreen, '<DevscreensButton />')) {
       patching.insertInFile(launchScreen, '</ScrollView>', '<DevscreensButton />', false)
