+++ packages/ignite-dev-screens/templates/APITestingScreen.ejs
@@ -2,11 +2,11 @@
 
 import React from 'react'
 import { ScrollView, View, Text, TouchableOpacity, Image } from 'react-native'
-import { Metrics, Images } from '../Themes'
-import FullButton from '../Components/FullButton'
+import { Metrics, Images } from '../../App/Themes'
+import FullButton from '../../App/Components/FullButton'
 
 // For API
-import API from '../Services/Api'
+import API from '../../App/Services/Api'
 import FJSON from 'format-json'
 
 // Styles
