+++ packages/ignite-dev-screens/templates/APITestingScreenStyle.ejs
@@ -1,7 +1,7 @@
 // @flow
 
 import { StyleSheet } from 'react-native'
-import { ApplicationStyles } from '../../Themes/'
+import { ApplicationStyles } from '../../../App/Themes/'
 
 export default StyleSheet.create({
   ...ApplicationStyles.screen
