+++ packages/ignite-dev-screens/templates/ComponentExamplesScreen.ejs
@@ -3,11 +3,11 @@
 // An All Components Screen is a great way to dev and quick-test components
 import React from 'react'
 import { Platform, View, ScrollView, Text, Image } from 'react-native'
-import { Images } from '../Themes'
-import styles from './Styles/AllComponentsScreenStyle'
+import { Images } from '../../App/Themes'
+import styles from './Styles/ComponentExamplesScreenStyle'
 
 // Examples Render Engine
-import ExamplesRegistry from '../Services/ExamplesRegistry'
+import ExamplesRegistry from '../../App/Services/ExamplesRegistry'
 
 class AllComponentsScreen extends React.Component {
