+++ packages/ignite-dev-screens/templates/ComponentExamplesScreenStyle.ejs
@@ -1,7 +1,7 @@
 // @flow
 
 import { StyleSheet } from 'react-native'
-import { ApplicationStyles } from '../../Themes/'
+import { ApplicationStyles } from '../../../App/Themes/'
 
 export default StyleSheet.create({
   ...ApplicationStyles.screen
