+++ packages/ignite-dev-screens/templates/DeviceInfoScreen.ejs
@@ -4,7 +4,7 @@
 import React from 'react'
 import { View, ScrollView, Text, Image, NetInfo } from 'react-native'
 import DeviceInfo from 'react-native-device-info'
-import { Metrics, Images } from '../Themes'
+import { Metrics, Images } from '../../App/Themes'
 import styles from './Styles/DeviceInfoScreenStyle'
 
 const HARDWARE_DATA = [
