+++ packages/ignite-dev-screens/templates/DeviceInfoScreenStyle.ejs
@@ -1,7 +1,7 @@
 // @flow
 
 import { StyleSheet } from 'react-native'
-import { Colors, Metrics, Fonts, ApplicationStyles } from '../../Themes/'
+import { Colors, Metrics, Fonts, ApplicationStyles } from '../../../App/Themes/'
 
 export default StyleSheet.create({
   ...ApplicationStyles.screen,
