+++ packages/ignite-dev-screens/templates/PluginExamplesScreen.ejs
@@ -4,13 +4,13 @@ import React, { PropTypes } from 'react'
 import { View, ScrollView, Text, TouchableOpacity, Image } from 'react-native'
 import { connect } from 'react-redux'
 import { Actions as NavigationActions } from 'react-native-router-flux'
-import { Colors, Images, Metrics } from '../Themes'
+import { Colors, Images, Metrics } from '../../App/Themes'
 
 // Examples Render Engine
-import ExamplesRegistry from '../Services/ExamplesRegistry'
+import ExamplesRegistry from '../../App/Services/ExamplesRegistry'
 
 // Styles
-import styles from './Styles/UsageExamplesScreenStyle'
+import styles from './Styles/PluginExamplesScreenStyle'
 
 class PluginExamplesScreen extends React.Component {
