+++ packages/ignite-dev-screens/templates/PluginExamplesScreenStyle.ejs
@@ -1,7 +1,7 @@
 // @flow
 
 import { StyleSheet } from 'react-native'
-import { Colors, Fonts, Metrics, ApplicationStyles } from '../../Themes/'
+import { Colors, Fonts, Metrics, ApplicationStyles } from '../../../App/Themes/'
 
 export default StyleSheet.create({
   ...ApplicationStyles.screen
