+++ packages/ignite-dev-screens/templates/PresentationScreen.ejs
@@ -2,8 +2,8 @@
 
 import React from 'react'
 import { ScrollView, Text, Image, View } from 'react-native'
-import { Images } from '../Themes'
-import RoundedButton from '../Components/RoundedButton'
+import { Images } from '../../App/Themes'
+import RoundedButton from '../../App/Components/RoundedButton'
 import { Actions as NavigationActions } from 'react-native-router-flux'
 
 // Styles
