+++ packages/ignite-dev-screens/templates/ThemeScreen.ejs
@@ -2,7 +2,7 @@
 
 import React from 'react'
 import { View, ScrollView, Text, Image } from 'react-native'
-import { Colors, Fonts, Images } from '../Themes'
+import { Colors, Fonts, Images } from '../../App/Themes'
 import R from 'ramda'
 
 // Styles
